Transformed input into correct type

+++ dataset.py
@@ -49,10 +49,10 @@ class SimulationDataset(Dataset):
         # Concatenate TF and F_expanded along a new dimension
         input_data = stack([TF, F_expanded], dim=2)  # input_data.shape = (501, 10, 2)
         
-        input_data = input_data.transpose(2, 0).transpose(1, 2)  # input_data.shape = (2, 501, 10)
+        input_data = input_data.transpose(2, 0).transpose(1, 2).to(torch.float32)  # input_data.shape = (2, 501, 10)
 
         # Extract observed momentum
-        Mtot = data['Mtot']
+        Mtot = data['Mtot'].to(torch.float32)
 
         # Apply the transformations if they exist
         if self.transform:
